refactor: renamed some classes
`ProjectBuilder` is now called `AppBuilder` while `RunContext` is now
`ProjectBuilder`.
