+++ src/main/java/dev/jbang/cli/Edit.java
@@ -59,14 +59,14 @@ public class Edit extends BaseCommand {
 	public Integer doCall() throws IOException {
 		scriptMixin.validate();
 
-		RunContext ctx = getRunContext();
-		final Project prj = ctx.forResource(scriptMixin.scriptOrFile);
+		ProjectBuilder pb = createProjectBuilder();
+		final Project prj = pb.build(scriptMixin.scriptOrFile);
 
 		if (prj.isJar() || prj.getMainSourceSet().getSources().isEmpty()) {
 			throw new ExitException(EXIT_INVALID_INPUT, "You can only edit source files");
 		}
 
-		Path project = createProjectForLinkedEdit(prj, ctx, false);
+		Path project = createProjectForLinkedEdit(prj, pb, false);
 		String projectPathString = Util.pathToString(project.toAbsolutePath());
 		// err.println(project.getAbsolutePath());
 
@@ -81,7 +81,7 @@ public class Edit extends BaseCommand {
 				// TODO only regenerate when dependencies changes.
 				info("Regenerating project.");
 				try {
-					createProjectForLinkedEdit(prj, RunContext.empty(), true);
+					createProjectForLinkedEdit(prj, ProjectBuilder.create(), true);
 				} catch (IOException e) {
 					throw new RuntimeException(e);
 				}
@@ -167,16 +167,16 @@ public class Edit extends BaseCommand {
 		return true;
 	}
 
-	RunContext getRunContext() {
-		RunContext ctx = new RunContext();
-		ctx.setProperties(dependencyInfoMixin.getProperties());
-		ctx.setAdditionalDependencies(dependencyInfoMixin.getDependencies());
-		ctx.setAdditionalRepositories(dependencyInfoMixin.getRepositories());
-		ctx.setAdditionalClasspaths(dependencyInfoMixin.getClasspaths());
-		ctx.setAdditionalSources(scriptMixin.sources);
-		ctx.setAdditionalResources(scriptMixin.resources);
-		ctx.setForceType(scriptMixin.forceType);
-		return ctx;
+	ProjectBuilder createProjectBuilder() {
+		return ProjectBuilder
+								.create()
+								.setProperties(dependencyInfoMixin.getProperties())
+								.additionalDependencies(dependencyInfoMixin.getDependencies())
+								.additionalRepositories(dependencyInfoMixin.getRepositories())
+								.additionalClasspaths(dependencyInfoMixin.getClasspaths())
+								.additionalSources(scriptMixin.sources)
+								.additionalResources(scriptMixin.resources)
+								.forceType(scriptMixin.forceType);
 	}
 
 	private static Optional<String> askEditor() throws IOException {
@@ -284,7 +284,7 @@ public class Edit extends BaseCommand {
 	}
 
 	/** Create Project to use for editing **/
-	Path createProjectForLinkedEdit(Project prj, RunContext ctx, boolean reload) throws IOException {
+	Path createProjectForLinkedEdit(Project prj, ProjectBuilder pb, boolean reload) throws IOException {
 		Path originalFile = prj.getResourceRef().getFile();
 
 		List<String> dependencies = prj.getMainSourceSet().getDependencies();
@@ -310,7 +310,7 @@ public class Edit extends BaseCommand {
 
 		for (ResourceRef sourceRef : prj.getMainSourceSet().getSources()) {
 			Path sfile = null;
-			Source src = ctx.createSource(sourceRef);
+			Source src = pb.createSource(sourceRef);
 			if (src.getJavaPackage().isPresent()) {
 				Path packageDir = srcDir.resolve(src.getJavaPackage().get().replace(".", File.separator));
 				Util.mkdirs(packageDir);
@@ -359,7 +359,7 @@ public class Edit extends BaseCommand {
 
 		renderTemplate(engine, depIds, fullClassName, baseName, resolvedDependencies, repositories,
 				templateName,
-				ctx.getArguments(),
+				pb.getArguments(),
 				destination);
 
 		// setup eclipse
@@ -367,14 +367,14 @@ public class Edit extends BaseCommand {
 		destination = tmpProjectDir.resolve(".classpath");
 		renderTemplate(engine, dependencies, fullClassName, baseName, resolvedDependencies, repositories,
 				templateName,
-				ctx.getArguments(),
+				pb.getArguments(),
 				destination);
 
 		templateName = ".qute.project";
 		destination = tmpProjectDir.resolve(".project");
 		renderTemplate(engine, dependencies, fullClassName, baseName, resolvedDependencies, repositories,
 				templateName,
-				ctx.getArguments(),
+				pb.getArguments(),
 				destination);
 
 		templateName = "main.qute.launch";
@@ -382,14 +382,14 @@ public class Edit extends BaseCommand {
 		destination.toFile().getParentFile().mkdirs();
 		renderTemplate(engine, dependencies, fullClassName, baseName, resolvedDependencies, repositories,
 				templateName,
-				ctx.getArguments(),
+				pb.getArguments(),
 				destination);
 
 		templateName = "main-port-4004.qute.launch";
 		destination = tmpProjectDir.resolve(".eclipse/" + baseName + "-port-4004.launch");
 		renderTemplate(engine, dependencies, fullClassName, baseName, resolvedDependencies, repositories,
 				templateName,
-				ctx.getArguments(),
+				pb.getArguments(),
 				destination);
 
 		// setup vscode
@@ -399,7 +399,7 @@ public class Edit extends BaseCommand {
 			destination.toFile().getParentFile().mkdirs();
 			renderTemplate(engine, dependencies, fullClassName, baseName, resolvedDependencies, repositories,
 					templateName,
-					ctx.getArguments(),
+					pb.getArguments(),
 					destination);
 		}
 
@@ -410,7 +410,7 @@ public class Edit extends BaseCommand {
 			destination.toFile().getParentFile().mkdirs();
 			renderTemplate(engine, dependencies, fullClassName, baseName, resolvedDependencies, repositories,
 					templateName,
-					ctx.getArguments(),
+					pb.getArguments(),
 					destination);
 		}
 
@@ -420,7 +420,7 @@ public class Edit extends BaseCommand {
 			destination.toFile().getParentFile().mkdirs();
 			renderTemplate(engine, dependencies, fullClassName, baseName, resolvedDependencies, repositories,
 					templateName,
-					ctx.getArguments(),
+					pb.getArguments(),
 					destination);
 		}
