+++ src/test/java/dev/jbang/cli/TestEdit.java
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.io.TempDir;
 
 import dev.jbang.*;
 import dev.jbang.source.Project;
-import dev.jbang.source.RunContext;
+import dev.jbang.source.ProjectBuilder;
 import dev.jbang.util.Util;
 
 import picocli.CommandLine;
@@ -46,10 +46,10 @@ public class TestEdit extends BaseTest {
 		JBang.getCommandLine().execute("init", s);
 		assertThat(new File(s).exists(), is(true));
 
-		RunContext ctx = RunContext.empty();
-		Project prj = ctx.forResource(s);
+		ProjectBuilder pb = ProjectBuilder.create();
+		Project prj = pb.build(s);
 
-		Path project = new Edit().createProjectForLinkedEdit(prj, ctx, false);
+		Path project = new Edit().createProjectForLinkedEdit(prj, pb, false);
 
 		assertThat(project.resolve("src").toFile(), FileMatchers.anExistingDirectory());
 		Path build = project.resolve("build.gradle");
@@ -86,10 +86,10 @@ public class TestEdit extends BaseTest {
 
 		Util.writeString(p, "//DEPS org.openjfx:javafx-graphics:11.0.2${bougus:}\n" + Util.readString(p));
 
-		RunContext ctx = RunContext.empty();
-		Project prj = ctx.forResource(s);
+		ProjectBuilder pb = ProjectBuilder.create();
+		Project prj = pb.build(s);
 
-		Path project = new Edit().createProjectForLinkedEdit(prj, ctx, false);
+		Path project = new Edit().createProjectForLinkedEdit(prj, pb, false);
 
 		Path gradle = project.resolve("build.gradle");
 		assert (Files.exists(gradle));
@@ -121,10 +121,10 @@ public class TestEdit extends BaseTest {
 				"//DEPS io.quarkus:quarkus-rest-client-reactive\n" +
 				"//DEPS io.quarkus:quarkus-rest-client-reactive-jackson\n" + Util.readString(p));
 
-		RunContext ctx = RunContext.empty();
-		Project prj = ctx.forResource(s);
+		ProjectBuilder pb = ProjectBuilder.create();
+		Project prj = pb.build(s);
 
-		Path project = new Edit().createProjectForLinkedEdit(prj, ctx, false);
+		Path project = new Edit().createProjectForLinkedEdit(prj, pb, false);
 
 		Path gradle = project.resolve("build.gradle");
 		assert (Files.exists(gradle));
@@ -154,10 +154,10 @@ public class TestEdit extends BaseTest {
 
 		Util.writeString(p, "//DEPS https://github.com/oldskoolsh/libvirt-schema/tree/0.0.2\n" + Util.readString(p));
 
-		RunContext ctx = RunContext.empty();
-		Project prj = ctx.forResource(s);
+		ProjectBuilder pb = ProjectBuilder.create();
+		Project prj = pb.build(s);
 
-		Path project = new Edit().createProjectForLinkedEdit(prj, ctx, false);
+		Path project = new Edit().createProjectForLinkedEdit(prj, pb, false);
 
 		Path gradle = project.resolve("build.gradle");
 		assert (Files.exists(gradle));
@@ -174,10 +174,10 @@ public class TestEdit extends BaseTest {
 		Path p = examplesTestFolder.resolve("one.java");
 		assertThat(p.toFile().exists(), is(true));
 
-		RunContext ctx = RunContext.empty();
-		Project prj = ctx.forResource(p.toString());
+		ProjectBuilder pb = ProjectBuilder.create();
+		Project prj = pb.build(p.toString());
 
-		Path project = new Edit().createProjectForLinkedEdit(prj, ctx, false);
+		Path project = new Edit().createProjectForLinkedEdit(prj, pb, false);
 
 		Path gradle = project.resolve("build.gradle");
 		assert (Files.exists(gradle));
@@ -199,10 +199,10 @@ public class TestEdit extends BaseTest {
 		assertThat(result, is(0));
 		assertThat(new File(s).exists(), is(true));
 
-		RunContext ctx = RunContext.empty();
-		Project prj = ctx.forResource(s);
+		ProjectBuilder pb = ProjectBuilder.create();
+		Project prj = pb.build(s);
 
-		Path project = new Edit().createProjectForLinkedEdit(prj, ctx, false);
+		Path project = new Edit().createProjectForLinkedEdit(prj, pb, false);
 
 		Path java = project.resolve("src/KubeExample.java");
 
@@ -217,10 +217,10 @@ public class TestEdit extends BaseTest {
 		Path p = examplesTestFolder.resolve("res/resource.java");
 		assertThat(p.toFile().exists(), is(true));
 
-		RunContext ctx = RunContext.empty();
-		Project prj = ctx.forResource(p.toString());
+		ProjectBuilder pb = ProjectBuilder.create();
+		Project prj = pb.build(p.toString());
 
-		Path project = new Edit().createProjectForLinkedEdit(prj, ctx, false);
+		Path project = new Edit().createProjectForLinkedEdit(prj, pb, false);
 
 		Path gradle = project.resolve("build.gradle");
 		assert (Files.exists(gradle));
